# Conflicts:
#	backend/src/main/java/com/ssafy/tnt/api/response/UserInfoGetRes.java
#	backend/src/main/java/com/ssafy/tnt/db/entity/User.java
#	frontend/src/components/main-header.vue
#	frontend/src/store/modules/accountStore.js
#	frontend/src/views/MainPage.vue

+++ backend/src/main/java/com/ssafy/tnt/api/service/NoticeService.java
@@ -11,6 +11,6 @@ public interface NoticeService {
 	public List<Notice> findAllNotice();
 	public Notice registNotice(NoticeDTO noticeDto);
 	public Notice modifyNotice(NoticeDTO noticeDto);
-	public Notice deleteNotice(int notice_no);
+	public Notice deleteNotice(Long notice_no);
 //	Notice findByNo(int no);
 }
